+++ lightly/transforms/torchvision_v2_compatibility.py
